requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
schedule>=1.1
gunicorn>=20.0